football-analysis-platform/apps/api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
pandas==2.2.0
scikit-learn==1.4.0
supabase>=2.9.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.6.0
sqlalchemy==2.0.25
scipy==1.12.0
numpy==1.26.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic[email]